refactor: rename source files extension to `.ts`
