+++ .eslintrc.json
@@ -11,7 +11,7 @@
   },
   "overrides": [
     {
-      "files": ["**/*.?([cm])ts"],
+      "files": ["**/*.ts"],
       "extends": ["plugin:@typescript-eslint/recommended"],
       "plugins": ["eslint-plugin-tsdoc"],
       "rules": {
