+++ action.yaml
@@ -10,4 +10,4 @@ inputs:
     required: true
 runs:
   using: node20
-  main: dist/action/index.mjs
+  main: dist/action/index.js
