--- src/index.mts
@@ -1,1 +1,0 @@
-export { pipxInstallAction } from "./action.mjs";
