+++ src/index.ts
@@ -1,0 +1,1 @@
+export { pipxInstallAction } from "./action.js";
