--- src/pipx/index.mts
@@ -1,11 +1,0 @@
-import { restorePackageCache, savePackageCache } from "./cache.mjs";
-import { ensurePath, getEnvironment } from "./environment.mjs";
-import { installPackage } from "./install.mjs";
-
-export default {
-  ensurePath,
-  getEnvironment,
-  installPackage,
-  restorePackageCache,
-  savePackageCache,
-};
