+++ src/pipx/index.ts
@@ -1,0 +1,11 @@
+import { restorePackageCache, savePackageCache } from "./cache.js";
+import { ensurePath, getEnvironment } from "./environment.js";
+import { installPackage } from "./install.js";
+
+export default {
+  ensurePath,
+  getEnvironment,
+  installPackage,
+  restorePackageCache,
+  savePackageCache,
+};
