+++ src/pipx/install.test.ts
@@ -25,7 +25,7 @@ describe("install Python packages", () => {
   });
 
   it("should successfully install a package", async () => {
-    const { installPackage } = await import("./install.mjs");
+    const { installPackage } = await import("./install.js");
 
     const prom = installPackage("ruff");
     await expect(prom).resolves.toBeUndefined();
@@ -34,7 +34,7 @@ describe("install Python packages", () => {
   });
 
   it("should fail to install an invalid package", async () => {
-    const { installPackage } = await import("./install.mjs");
+    const { installPackage } = await import("./install.js");
 
     const prom = installPackage("invalid-pkg");
     await expect(prom).rejects.toThrow("Failed to install invalid-pkg");
